+++ Sources/Pippin/Adapters/COSTouchVisualizer/COSTouchVisualizerAdapter.swift
@@ -1,0 +1,43 @@
+//
+//  COSTouchVisualizerAdapter.swift
+//  Pippin
+//
+//  Created by Andrew McKnight on 9/14/18.
+//
+
+import COSTouchVisualizer
+import Foundation
+
+/// - Note: `COSTouchVisualizer` requires using its `UIWindow` subclass to construct the main view hierarchy, so this class instantiates and maintains a reference to it, and upon calling `installViews()` calls `makeKeyAndVisible()` on it, thus presenting your app as it normally would be in its `UIApplicationDelegate`.
+public class COSTouchVisualizerAdapter: NSObject {
+    private let window: UIWindow
+    
+    public init(rootViewController: UIViewController) {
+        let touchWindow = COSTouchVisualizerWindow(frame: UIScreen.main.bounds)
+        self.window = touchWindow
+        
+        super.init()
+        
+        touchWindow.rootViewController = rootViewController
+        touchWindow.touchVisualizerWindowDelegate = self
+        touchWindow.stationaryMorphEnabled = false
+    }
+}
+
+// MARK: TouchVisualization
+extension COSTouchVisualizerAdapter: TouchVisualization {
+    public func installViews() {
+        window.makeKeyAndVisible()
+    }
+}
+
+// MARK: COSTouchVisualizerWindowDelegate
+extension COSTouchVisualizerAdapter: COSTouchVisualizerWindowDelegate {
+    public func touchVisualizerWindowShouldShowFingertip(_ window: COSTouchVisualizerWindow) -> Bool {
+        return true
+    }
+    
+    public func touchVisualizerWindowShouldAlwaysShowFingertip(_ window: COSTouchVisualizerWindow!) -> Bool {
+        return true
+    }
+}