+++ Sources/Pippin/Core/UI/TouchVisualization.swift
@@ -1,0 +1,12 @@
+//
+//  TouchVisualization.swift
+//  Pippin
+//
+//  Created by Andrew McKnight on 9/14/18.
+//
+
+import Foundation
+
+public protocol TouchVisualization {
+    func installViews()
+}